+++ src/IntervalAlgebra/PairedInterval.hs
@@ -39,6 +39,9 @@ instance (Ord a) => Intervallic (PairedInterval b) a where
     getInterval (PairedInterval x)        = fst x
     setInterval (PairedInterval (x, y)) i = PairedInterval (i, y)
 
+instance Functor (PairedInterval b) where
+    fmap f (PairedInterval (x, y)) = PairedInterval (fmap f x, y)
+
 instance Bifunctor PairedInterval where
     bimap f g (PairedInterval (x, y)) = PairedInterval (fmap g x, f y)