+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for interval-algebra
 
+## 0.10.0
+
+* Adds `diffFromBegin` (`diffFromEnd`) functions (not totally satisfied with these names) which change the reference point of the interval in the second argument by the difference from the `begin` (`end`) of the interval in the first argument.
+* Adds a `Functor` instance for `PairedInterval b`s, which maps an `PairedInterval c a` to `PairedInterval c b`. That is, `fmap` acts on the interval type.
+
 ## 0.9.0
 
 * Fixes bug in `gapsWithin` introduced in last version.